requirements.txt
discord.py
langchain-google-genai
langgraph
langchain-core
requests
python-dotenv